WorkingWithPromptFlowSDKinVSCode/05-flexflow/requirements.txt
python-dotenv==1.0.1
promptflow
